added pagination for users collection and refactored elastic and local controllers directories
